+++ Runtime/Components/BaseAttribute.cs
@@ -5,33 +5,33 @@ namespace LittleBit.Modules.Description.Components
     [Serializable]
     public class BaseAttribute
     {
-        protected float _baseValue, _baseMultiplier;
+        protected double _baseValue, _baseMultiplier;
         protected bool _isDirty = true;
         
-        public BaseAttribute(float baseValue, float baseMultiplier)
+        public BaseAttribute(double baseValue, double baseMultiplier)
         {
             _baseValue = baseValue;
             _baseMultiplier = baseMultiplier;
         }
 
-        public float GetBaseValue()
+        public double GetBaseValue()
         {
             return _baseValue;
         }
 
-        public float GetBaseMultiplier()
+        public double GetBaseMultiplier()
         {
             return _baseMultiplier;
         }
         
         
-        public void SetBaseValue(float value)
+        public void SetBaseValue(double value)
         {
             _baseValue = value;
             _isDirty = true;
         }
 
-        public void SetBaseMultiplier(float value)
+        public void SetBaseMultiplier(double value)
         {
             _baseMultiplier = value;
             _isDirty = true;
